Move importer actions into separate files
