+++ tests/Unit/IO/Importers/Test_VolumeDataReaderActions.cpp
@@ -22,9 +22,10 @@
 #include "IO/H5/AccessType.hpp"
 #include "IO/H5/File.hpp"
 #include "IO/H5/VolumeData.hpp"
-#include "IO/Importers/ElementActions.hpp"
+#include "IO/Importers/Actions/ReadVolumeData.hpp"
+#include "IO/Importers/Actions/ReceiveVolumeData.hpp"
+#include "IO/Importers/Actions/RegisterWithElementDataReader.hpp"
 #include "IO/Importers/ElementDataReader.hpp"
-#include "IO/Importers/ElementDataReaderActions.hpp"
 #include "IO/Importers/Tags.hpp"
 #include "IO/Observer/ArrayComponentId.hpp"
 #include "NumericalAlgorithms/Spectral/Spectral.hpp"
